Instruction Lists
Closes #1. The main form in which each bot tracks its path is changed. Each bot now has a list of a new Instruction class. Each instruction says what node to move to, at what time should the bot start moving to the node and the distance to the node.

This, along with an additional bool, allows the bot to make sure that they move for a scheduleItem when the scheduleItem is happening soon, and not  as soon as it is able to move. This means that bot will hang around the node that their last ScheduleItem happened, and not where their next ScheduleItem will be.